+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
+All notable changes to this project will the documented in this file.
+#### Added
+- Support for custom validation rules; which include ValidateName, ValidateEmail, ValidatePhoneNumber, ValidateAddress, ValidateCity, ValidateState, ValidateZipCode
 
-* TODO: Describe initial release.
+## 0.1.0
+
+#### Added
+- Fixed new bugs